make more flexible
