+++ applications/CoSimulationApplication/tests/test_mok_fsi.py
@@ -61,7 +61,7 @@ class TestMokFSI(co_simulation_test_case.CoSimulationTestCase):
             self.__RemoveOutputFromCFD() # comment to get output
             self.__AddTestingToCFD()
             self.__DumpUpdatedCFDSettings()
-            self._runTestWithExternal([GetPython3Command(), "structural_mechanics_analysis_with_co_sim_io.py", ext_parameter_file_name])
+            self._runTestWithExternal([GetPython3Command(), "testing_structural_mechanics_analysis_with_co_sim_io.py", ext_parameter_file_name])
 
     def test_mok_fsi_mvqn_external_structure_remote_controlled(self):
         self.accelerator_type = "mvqn"
@@ -91,7 +91,7 @@ class TestMokFSI(co_simulation_test_case.CoSimulationTestCase):
             self.cfd_parameters["problem_data"]["parallel_type"].SetString("MPI")
             self.__DumpUpdatedCFDSettings()
             num_procs = KM.Testing.GetDefaultDataCommunicator().Size()
-            self._runTestWithExternal(["mpiexec", "-np", str(num_procs), GetPython3Command(), "structural_mechanics_analysis_with_co_sim_io.py", "--using-mpi", ext_parameter_file_name])
+            self._runTestWithExternal(["mpiexec", "-np", str(num_procs), GetPython3Command(), "testing_structural_mechanics_analysis_with_co_sim_io.py", "--using-mpi", ext_parameter_file_name])
 
 
     def __ManipulateCFDSettings(self):
